move ExecuteCommand to Commands
