+++ project.properties
@@ -1,1 +1,1 @@
-releaseVersion=2.0.6
+releaseVersion=2.0.7
